requirements.txt
argcomplete==3.1.2
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
boto3==1.28.43
botocore==1.31.43
certifi==2023.7.22
cffi==1.15.1
cfn-flip==1.3.0
charset-normalizer==3.2.0
click==8.1.7
constantly==15.1.0
cryptography==41.0.2
dj-database-url==2.1.0
Django==4.2.3
django-appconf==1.0.5
django-db-connection-pool==1.2.4
django-guest-user==0.5.4
django-imagekit==4.1.0
django-storages==1.14
django-widget-tweaks==1.5.0
durationpy==0.5
greenlet==2.0.2
gunicorn==21.2.0
hjson==3.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.3
msgpack==1.0.5
packaging==23.1
pilkit==2.0
Pillow==9.0.0
placebo==0.9.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.1.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
PyYAML==6.0.1
rcssmin==1.1.1
requests==2.31.0
rjsmin==1.2.1
s3transfer==0.6.2
service-identity==23.1.0
six==1.16.0
sorl-thumbnail==12.9.0
SQLAlchemy==2.0.19
sqlparams==5.1.0
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.1
troposphere==4.4.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.7
zappa==0.58.0
zope.interface==6.0
